refactor: Split postgres_provisioner.go into logical modules
Break the 1408-line postgres_provisioner.go into four focused files,
each under 600 lines, to improve navigability and maintainability:

- postgres_provisioner.go (602 lines): Core struct, constructor,
  provisioning/deprovisioning lifecycle, circuit breaker orchestration
- migration_runner.go (451 lines): Migration reading, execution,
  SQL processing, reconciliation logic
- provisioner_persistence.go (154 lines): Database entity, GORM
  persistence operations, entity/status conversion
- provisioner_helpers.go (152 lines): Connection management, schema
  DDL operations, SQL utilities, error classification

No behavioral changes. All existing tests pass unchanged.
